fix: exclude dsas_js.tcz from upgrade

+++ make.sh
@@ -797,6 +797,7 @@ upgrade)
     pkg_file=${pkg_file%-doc}.pkg
     [ -f "$pkg_file" ] && continue   # Locally built package
     [ "$_file" = "dsastestfiles.tcz" ] && continue  # Locally built file
+    [ "$_file" = "dsas_js.tcz" ] && continue  # Locally built file
     [ "$_file" = "dsaswebdriver.tcz" ] && { msg "Removing $_file"; rm -f "$file"; continue; } # Remove to force rebuild
     [ "$_file" = "firefox.tcz" ] && { msg "Removing $_file"; rm -f "$file"; continue; } # Remove to force a rebuild
     read -r hash < "$file.md5.txt"
